Add evaluate_ensemble support to EvaluateRepositoryCollection
